requirements.txt
alpaca-py>=0.7
pandas>=1.5